add simple unit test
