+++ opengrok-tools/src/test/python/test_xml.py
@@ -1,0 +1,38 @@
+#!/usr/bin/env python3
+
+#
+# CDDL HEADER START
+#
+# The contents of this file are subject to the terms of the
+# Common Development and Distribution License (the "License").
+# You may not use this file except in compliance with the License.
+#
+# See LICENSE.txt included in this distribution for the specific
+# language governing permissions and limitations under the License.
+#
+# When distributing Covered Code, include this CDDL HEADER in each
+# file and include the License file at LICENSE.txt.
+# If applicable, add the following below this CDDL HEADER, with the
+# fields enclosed by brackets "[]" replaced with your own identifying
+# information: Portions Copyright [yyyy] [name of copyright owner]
+#
+# CDDL HEADER END
+#
+
+#
+# Copyright (c) 2019, Oracle and/or its affiliates. All rights reserved.
+#
+
+import os
+
+from opengrok_tools.utils.xml import insert_file
+
+
+def test_xml_insert():
+    dir_path = os.path.dirname(os.path.realpath(__file__))
+
+    with open(os.path.join(dir_path, "web.xml")) as base_xml:
+        out = insert_file(base_xml.read(),
+                          os.path.join(dir_path, "insert.xml"))
+        with open(os.path.join(dir_path, "new.xml")) as expected_xml:
+            assert out == expected_xml.read()
